+++ azext_iot/tests/digitaltwins/__init__.py
@@ -62,7 +62,7 @@ EP_EVENTGRID_TOPIC = settings.env.azext_dt_ep_eventgrid_topic or ("test-egt-" + 
 ADX_CLUSTER = settings.env.azext_dt_adx_cluster or ("testadxc" + generate_generic_id()[:4])
 ADX_DATABASE = settings.env.azext_dt_adx_database or ("testadxd" + generate_generic_id()[:4])
 ADX_RG = settings.env.azext_dt_adx_rg or settings.env.azext_iot_testrg
-MAX_MODEL_PER_BATCH = settings.env.azext_dt_max_models_per_batch
+MAX_MODELS_PER_BATCH = settings.env.azext_dt_max_models_per_batch
 
 STORAGE_ACCOUNT = settings.env.azext_iot_teststorageaccount or "testadtstore" + generate_generic_id()[:4]
 STORAGE_CONTAINER = settings.env.azext_iot_teststoragecontainer or DEFAULT_CONTAINER